test(ctl): split the schema load error tests by responsibility
The single test module covering the rendering of `schema load` errors is
split into one file per concern under tests/unit/ctl/schema/: parsing of
the schema-level message, locating an error in the submitted schema,
rendering of field-level errors, and rendering of schema-level errors
together with the server contract checks. The offline validation and
warning tests join test_schema_app.py, which already covers that path.
Builders and schema payloads shared across those files move to
tests/helpers/schema_load_errors.py.
